# Optimize Client.Deactivate by removing N+1 query