--- a/test/complex/mongo_client_test.go
+++ b/test/complex/mongo_client_test.go
@@ -72,8 +72,8 @@ func TestClientWithShardedDB(t *testing.T) {
 		testcases.RunFindDocInfoTest(t, cli, dummyProjectID)
 	})
 
-	t.Run("RunFindDocInfosByKeys test", func(t *testing.T) {
-		testcases.RunFindDocInfosByKeysTest(t, cli, dummyProjectID)
+	t.Run("RunFindDocInfosByKeysAndIDs test", func(t *testing.T) {
+		testcases.RunFindDocInfosByKeysAndIDsTest(t, cli, dummyProjectID)
 	})
 
 	t.Run("RunFindDocInfosByQuery test", func(t *testing.T) {