<!--  Thanks for sending a pull request! -->

**What this PR does / why we need it**:
 In `Yorkie.DeactivateClient`, firstly it finds attached or attaching documents to detach via DB query.
Previously, the number of query is N+1 (1 for `FindClientInfo`, N for `FindDocInfoByRefKey`) to find it.
Execute `N * FindDocInfoByRefKey` can be changed to batch query `FindDocInfosByDocIDs` in both mongoDB and memDB to reduce execution time.

**Which issue(s) this PR fixes**:

<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Fixes #

**Special notes for your reviewer**:
 Let me share benchmark result for mongoDB case. You can refer `deactivate_bench_test.go` for detailed scenario.
This change reduces about **10%** of `Client.Deactivate` execution time in all number of attached documents, while memory consumption and number of allocation is almost same.

- Small case (attached document is 1 to 20)

| Case                                | Before (ms) | After (ms) | % (after/before) |
|-------------------------------------|-------------|------------|------------------|
| deactivate_with_1_total_1_attached  | 2.86        | 2.58       | 90.11%           |
| deactivate_with_2_total_2_attached  | 4.65        | 4.55       | 97.93%           |
| deactivate_with_3_total_3_attached  | 6.14        | 5.90       | 96.11%           |
| deactivate_with_4_total_4_attached  | 8.73        | 7.90       | 90.50%           |
| deactivate_with_5_total_5_attached  | 11.53       | 9.14       | 79.27%           |
| deactivate_with_6_total_6_attached  | 12.27       | 13.72      | 111.80%          |
| deactivate_with_7_total_7_attached  | 12.84       | 12.39      | 96.45%           |
| deactivate_with_8_total_8_attached  | 17.66       | 13.89      | 78.64%           |
| deactivate_with_9_total_9_attached  | 17.27       | 14.34      | 83.02%           |
| deactivate_with_10_total_10_attached | 19.90      | 17.01      | 85.46%           |
| deactivate_with_11_total_11_attached | 19.34      | 20.38      | 105.41%          |
| deactivate_with_12_total_12_attached | 22.46      | 21.45      | 95.53%           |
| deactivate_with_13_total_13_attached | 22.46      | 23.53      | 104.76%          |
| deactivate_with_14_total_14_attached | 23.85      | 24.97      | 104.69%          |
| deactivate_with_15_total_15_attached | 29.06      | 29.00      | 99.82%           |
| deactivate_with_16_total_16_attached | 29.76      | 31.47      | 105.74%          |
| deactivate_with_17_total_17_attached | 31.52      | 28.24      | 89.59%           |
| deactivate_with_18_total_18_attached | 32.24      | 29.85      | 92.59%           |
| deactivate_with_19_total_19_attached | 36.86      | 30.76      | 83.44%           |
| deactivate_with_20_total_20_attached | 39.46      | 34.26      | 86.83%           |

- Large case (x-axis is attached documents in client, measured from 10 to 1000 in step of 10)

**Execution Time**
<img width="704" height="422" alt="image" src="https://github.com/user-attachments/assets/cbf86d26-14b4-4b9f-a965-be9bfc8696df" />

**Memory Usage**
<img width="704" height="422" alt="image" src="https://github.com/user-attachments/assets/abfcc697-1257-47c2-9bc2-e714eea2cd81" />

**Number of Allocation**
<img width="704" height="422" alt="image" src="https://github.com/user-attachments/assets/ad126c45-864b-42b3-8535-0c613d7e0897" />

**Does this PR introduce a user-facing change?**:

<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".
-->

```release-note
NONE
```

**Additional documentation**:

<!--
This section can be blank if this pull request does not require a release note.

Please use the following format for linking documentation:
- [Usage]: <link>
- [Other doc]: <link>
-->

```docs

```

**Checklist**:

- [X] Added relevant tests or not required
- [X] Addressed and resolved all CodeRabbit review comments
- [X] Didn't break anything


<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

* **New Features**
  * Batch document retrieval by IDs for faster multi-document operations.
* **Refactor**
  * Client deactivation now uses batched lookups with validation of fetched results to improve performance and reliability.
* **Tests**
  * Expanded tests to cover retrieval by keys and IDs (including empty/not‑found cases) and updated test labels.
  * Added deactivation benchmarks to track performance.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->